refactor(StoryError): extract ErrorTextFormatter
This is going to help in reducing duplication amoung other types
of diagnostic reporting like deprecations.
